[#25] - Improve code by nested APIQuestion inside APISurvey and APIAnswer inside APIQuestion
