+++ Surveys/Sources/Data/NetworkAPI/Models/APISurvey.swift
@@ -27,3 +27,34 @@ extension APISurvey {
         case questionList = "questions"
     }
 }
+
+extension APISurvey {
+
+    struct APIQuestion: Question, Codable, Equatable {
+
+        struct APIAnswer: Answer, Codable, Equatable {
+
+            let id: String
+            let type: String
+            let displayOrder: Int
+            let text: String?
+            let inputMaskPlaceholder: String?
+        }
+
+        let id: String
+        let type: String
+        let displayOrder: Int
+        let displayType: String
+        let text: String
+        let pick: PickType
+        let coverImageUrl: String
+        var answerList: [APIAnswer]?
+        var answers: [Answer]? { answerList }
+
+        enum CodingKeys: String, CodingKey {
+
+            case id, text, pick, type, displayOrder, displayType, coverImageUrl
+            case answerList = "answers"
+        }
+    }
+}
